Refactor concrete state classes.
Now the base class SimulationDataContainer no longer
sets up default fields, nor does it have accessor methods
like pressure() or saturation(). Instead, that is up to
the concrete state class to provide.

The accessors have been made with the hope that performance
will be as good as before. Note that it is necessary to
reseat the pointers after copy construction or assignment.